requirements.txt
requests
flask
flask-restful
flask-jwt-extended
flask-migrate
flask-cors
flask-swagger-ui
flasgger
coverage
python-decouple
Flask==2.3.2
Flask-JWT-Extended==4.4.4
werkzeug==2.3.0